Extract event queue to new file
